docs: prepare v7.2 release gate
